Hubbub/EtriControlAgent/requirements.txt
mysql-connector==2.2.9
pip==19.2.3
setuptools==40.8.0
